+++ gst/nnstreamer/nnstreamer.c
@@ -54,8 +54,8 @@
 #include "tensor_filter/tensor_filter.h"
 #include "tensor_merge/gsttensormerge.h"
 #include "tensor_mux/gsttensormux.h"
-#include "tensor_reposink/tensor_reposink.h"
-#include "tensor_reposrc/tensor_reposrc.h"
+#include "tensor_repo/tensor_reposink.h"
+#include "tensor_repo/tensor_reposrc.h"
 #include "tensor_sink/tensor_sink.h"
 #include "tensor_source/tensor_src_iio.h"
 #include "tensor_split/gsttensorsplit.h"
