+++ gst/nnstreamer/tensor_repo/meson.build
@@ -1,7 +1,9 @@
-tensor_reposrc_sources = [
+tensor_repo_sources = [
+  'tensor_repo.c',
+  'tensor_reposink.c',
   'tensor_reposrc.c'
 ]
 
-foreach s : tensor_reposrc_sources
+foreach s : tensor_repo_sources
   nnstreamer_sources += join_paths(meson.current_source_dir(), s)
 endforeach
