+++ gst/nnstreamer/tensor_repo/tensor_repo.c
