--- gst/nnstreamer/tensor_reposink/meson.build
@@ -1,7 +1,0 @@
-tensor_reposink_sources = [
-  'tensor_reposink.c'
-]
-
-foreach s : tensor_reposink_sources
-  nnstreamer_sources += join_paths(meson.current_source_dir(), s)
-endforeach
